feat: add one filters component
