+++ src/pages/Catalog/Catalog.tsx
@@ -2,9 +2,6 @@ import React, { useContext } from "react";
 import "./Catalog.css";
 import db from "../../assets/db.json";
 import Card from "../../components/Card/Card";
-import Categories from "../../components/Category-filter/Category-filter";
-import Brand from "../../components/Brand-filter/Brand-filter";
-import SortOptions from "../../components/Sort-filters/Sort-filter";
 import { TotalContext } from "../../totalContext";
 import { ICountType } from "../../types";
 
@@ -23,7 +20,7 @@ const Catalog = () => {
     images: string[];
   };
 
-  const {totalPrice, setTotalPrice} = useContext(TotalContext);
+  const { totalPrice, setTotalPrice } = useContext(TotalContext);
 
   const basketProducts = JSON.parse(
     localStorage.getItem("basketProducts") || `[]`
@@ -37,12 +34,15 @@ const Catalog = () => {
   };
 
   const addToCart = (object: dataProps) => {
-    const total = JSON.parse(localStorage.getItem('total')!);
+    const total = JSON.parse(localStorage.getItem("total")!);
     setTotalPrice(total.price + object.price);
-    localStorage.setItem('total', JSON.stringify({
-      count: total.count + 1,
-      price: total.price + object.price
-    }));
+    localStorage.setItem(
+      "total",
+      JSON.stringify({
+        count: total.count + 1,
+        price: total.price + object.price,
+      })
+    );
     setTotalPrice(totalPrice + object.price);
     isAdded(object);
     basketProducts.push(object);
@@ -51,16 +51,27 @@ const Catalog = () => {
 
   const removeFromCart = (object: dataProps) => {
     setTotalPrice(totalPrice - object.price);
-    const isCounts = JSON.parse(localStorage.getItem('counts')!) !== null ? true : false;
-    const counts = JSON.parse(localStorage.getItem('counts')!);
-    const total = JSON.parse(localStorage.getItem('total')!);
-    setTotalPrice(total.price - object.price * (isCounts ? counts[`${object.id}`] : 1));
-    localStorage.setItem('total', JSON.stringify({
-      count: total.count - 1,
-      price: total.price - object.price * (isCounts ? counts[`${object.id}`] : 1)
-    }));
-    const indexOfObj = basketProducts.findIndex((item: dataProps) => item.id === object.id);
-    basketProducts.sort().splice(indexOfObj, isCounts ? counts[`${object.id}`] : 1);
+    const isCounts =
+      JSON.parse(localStorage.getItem("counts")!) !== null ? true : false;
+    const counts = JSON.parse(localStorage.getItem("counts")!);
+    const total = JSON.parse(localStorage.getItem("total")!);
+    setTotalPrice(
+      total.price - object.price * (isCounts ? counts[`${object.id}`] : 1)
+    );
+    localStorage.setItem(
+      "total",
+      JSON.stringify({
+        count: total.count - 1,
+        price:
+          total.price - object.price * (isCounts ? counts[`${object.id}`] : 1),
+      })
+    );
+    const indexOfObj = basketProducts.findIndex(
+      (item: dataProps) => item.id === object.id
+    );
+    basketProducts
+      .sort()
+      .splice(indexOfObj, isCounts ? counts[`${object.id}`] : 1);
     localStorage.setItem("basketProducts", JSON.stringify(basketProducts));
   };
 
@@ -71,6 +82,93 @@ const Catalog = () => {
   ): void => {
     setSearchValue(event.target.value);
   };
+  //filters
+  //filters category
+  const searchCategory = [
+    "smartphones",
+    "laptops",
+    "fragrances",
+    "skincare",
+    "groceries",
+    "home-decoration",
+    "furniture",
+    "tops",
+    "womens-dresses",
+    "womens-shoes",
+    "mens-shirts",
+    "mens-shoes",
+    "mens-watches",
+    "mens-bags",
+    "womens-jewelerry",
+    "sunglasses",
+    "automotive",
+    "automotive",
+    "motorcycle",
+    "lighting",
+  ];
+  //brand filters
+  const brandCategory = [
+    "Apple",
+    "Samsung",
+    "OPPO",
+    "Huawei",
+    "Microsoft Surface",
+    "Infinix",
+    "HP Pavilion",
+    "Impression of Acqua Di Gio",
+    "Royal_Mirage",
+    "Fog Scent Xpressio",
+    "Al Munakh",
+    "Lord - Al-Rehab",
+    "L'Oreal Paris",
+    "Hemani Tea",
+    "Dermive",
+    "ROREC White Rice",
+    "Fair & Clear",
+    "Saaf & Khaas",
+    "Bake Parlor Big",
+    "Baking Food Items",
+    "fauji",
+    "Dry Rose",
+    "Boho Decor",
+    "Flying Wooden",
+    "LED Lights",
+    "luxury palace",
+    "Golden",
+    "Furniture Bed Set",
+    "Ratttan Outdoor",
+    "Kitchen Shelf",
+    "Multi Purpose",
+    "AmnaMart",
+    "Professional Wear",
+    "Soft Cotton",
+    "Top Sweater",
+    "RED MICKY MOUSE..",
+    "Digital Printed",
+    "Ghazi Fabric",
+    "IELGY",
+    "IELGY fashion",
+    "Synthetic Leather",
+    "Sandals Flip Flops",
+    "Maasai Sandals",
+    "Arrivals Genuine",
+    "Vintage Apparel",
+    "FREE FIRE",
+    "The Warehouse",
+    "Sneakers",
+    "Rubber",
+    "Naviforce",
+    "SKMEI 9117",
+    "Strap Skeleton",
+    "Stainless",
+    "Eastern Watches",
+    "Luxury Digital",
+    "Watch Pearls",
+    "Bracelet",
+    "LouisWill",
+    "Copenhagen Luxe",
+    "Steal Frame",
+  ];
   return (
     <div className="catalog">
       <div className="catalog-container align-center">
@@ -81,11 +179,27 @@ const Catalog = () => {
                 <button className="reset__block-button">Reset Filters</button>
               </div>
               <div className="filters__block-title">Category</div>
-              <Categories />
+              <div className="category">
+                {searchCategory.map((value, i) => (
+                  <div key={i} className="category__checkbox">
+                    <input key={i} type="checkbox" />
+                    <label>{value}</label>
+                    <span>(5/5)</span>
+                  </div>
+                ))}
+              </div>
               <div className="brands">
                 <div className="filters__block-title">Brand</div>
                 <div className="brands__list">
-                  <Brand />
+                  <div className="brands__list-item">
+                    {brandCategory.map((value, i) => (
+                      <div key={i}>
+                        <input type="checkbox" />
+                        <label>{value}</label>
+                        <span>(0/1)</span>
+                      </div>
+                    ))}
+                  </div>
                 </div>
                 <div className="price__block">
                   <div className="filters__block-title">Price</div>
@@ -100,7 +214,24 @@ const Catalog = () => {
                 <div className="search__title">
                   {searchValue ? `Search by: "${searchValue}"` : "All products"}
                 </div>
-                <SortOptions />
+                <div className="sort__bar">
+                  <select name="" id="" className="sort__bar-list">
+                    <option
+                      value="sort-title"
+                      disabled
+                      selected
+                      className="sort__bar-name"
+                    >
+                      Sort options:
+                    </option>
+                    <option value="price-ASC">Sort by price ASC</option>
+                    <option value="price-DESC">Sort by price DESC</option>
+                    <option value="rating-ASC">Sort by raiting ASC</option>
+                    <option value="rating-DESC">Sort by raiting DESC</option>
+                    <option value="discount-ASC">Sort by discount ASC</option>
+                    <option value="discount-DESC">Sort by descount DESC</option>
+                  </select>
+                </div>
                 <div className="found__title">Found:1</div>
                 <div className="search__block">
                   <input
